Implemented manual entering of text
The user can now enter text in manually. They can choose to write it to a new file, in which they can name the file, or if they don't want to it will be saved in a temporary file, which will be overwritten the next time the enter text manually.
Some improvements to Settings function, although there is one weird bug remaining.